chore: Remove duplicate migrations. Local-only telescope tables.

+++ database/migrations/2025_07_03_110517_create_telescope_entries_table.php
@@ -19,6 +19,11 @@ return new class extends Migration
      */
     public function up(): void
     {
+        // Prevent unnecessary database overhead and storage usage in production.
+        if (!app()->environment('local')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->create('telescope_entries', function (Blueprint $table) {
@@ -61,6 +66,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (!app()->environment('local')) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->dropIfExists('telescope_entries_tags');
